Added Mac Test support for vkdispatch vs vkfft convolution testing

+++ fft_2d/run_test.sh
@@ -15,18 +15,24 @@ BATCH_SIZE=$3
 REPEATS=$4
 ARCH=$5
 
-# /usr/local/cuda/bin/nvcc ../fft_cufft.cu -o fft_cufft.exec -lcufft
-
 echo "Running performance tests with the following parameters:"
 echo "Data Size: $DATA_SIZE"
 echo "Iteration Count: $ITER_COUNT"
 echo "Batch Size: $BATCH_SIZE"
 echo "Repeats: $REPEATS"
 
-echo "Running cuFFT FFT..."
-$NVCC -O2 -std=c++17 ../cufft_test.cu -gencode arch=compute_${ARCH},code=sm_${ARCH} -lcufft -lculibos -o cufft_test.exec
-./cufft_test.exec $DATA_SIZE $ITER_COUNT $BATCH_SIZE $REPEATS
-rm cufft_test.exec
+if [[ $(uname) != "Darwin" ]]; then
+    echo "Running cuFFT FFT..."
+    $NVCC -O2 -std=c++17 ../cufft_test.cu -gencode arch=compute_${ARCH},code=sm_${ARCH} -lcufft -lculibos -o cufft_test.exec
+    ./cufft_test.exec $DATA_SIZE $ITER_COUNT $BATCH_SIZE $REPEATS
+    rm cufft_test.exec
+
+    echo "Running PyTorch FFT..."
+    python3 ../torch_test.py $DATA_SIZE $ITER_COUNT $BATCH_SIZE $REPEATS
+
+    echo "Running ZipFFT FFT..."
+    python3 ../zipfft_test.py $DATA_SIZE $ITER_COUNT $BATCH_SIZE $REPEATS
+fi
 
 echo "Running Vkdispatch FFT..."
 python3 ../vkdispatch_test.py $DATA_SIZE $ITER_COUNT $BATCH_SIZE $REPEATS
@@ -34,8 +40,3 @@ python3 ../vkdispatch_test.py $DATA_SIZE $ITER_COUNT $BATCH_SIZE $REPEATS
 echo "Running VKFFT FFT..."
 python3 ../vkfft_test.py $DATA_SIZE $ITER_COUNT $BATCH_SIZE $REPEATS
 
-echo "Running PyTorch FFT..."
-python3 ../torch_test.py $DATA_SIZE $ITER_COUNT $BATCH_SIZE $REPEATS
-
-echo "Running ZipFFT FFT..."
-python3 ../zipfft_test.py $DATA_SIZE $ITER_COUNT $BATCH_SIZE $REPEATS
